+++ now/executor/indexer/elastic/elastic_indexer.py
@@ -150,16 +150,13 @@ class NOWElasticIndexer(Executor):
         """
         if not docs_map:
             return DocumentArray()
+        self.aggregate_embeddings(docs_map)
         preprocessed_docs_map = merge_subdocuments(docs_map, self.encoder_to_fields)
         es_docs = ESConverter.convert_doc_map_to_es(
             preprocessed_docs_map, self.index_name, self.encoder_to_fields
         )
-        try:
-            success, _ = bulk(self.es, es_docs)
-            self.es.indices.refresh(index=self.index_name)
-        except Exception as e:
-            self.logger.info(traceback.format_exc())
-            raise
+        success, _ = bulk(self.es, es_docs)
+        self.es.indices.refresh(index=self.index_name)
         if success:
             self.logger.info(
                 f'Inserted {success} documents into Elasticsearch index {self.index_name}'
@@ -199,6 +196,7 @@ class NOWElasticIndexer(Executor):
         """
         if not docs_map:
             return DocumentArray()
+        self.aggregate_embeddings(docs_map)
 
         # search_filter = parameters.get('filter', None)
         limit = parameters.get('limit', self.limit)
@@ -222,23 +220,20 @@ class NOWElasticIndexer(Executor):
             query_to_curated_ids=self.query_to_curated_ids,
         )
         for doc, query in es_queries:
-            try:
-                result = self.es.search(
-                    index=self.index_name,
-                    query=query,
-                    source=True,
-                    size=limit,
-                )['hits']['hits']
-                doc.matches = ESConverter.convert_es_results_to_matches(
-                    query_doc=doc,
-                    es_results=result,
-                    get_score_breakdown=get_score_breakdown,
-                    metric=self.metric,
-                    semantic_scores=self.default_semantic_scores,
-                )
-                doc.tags.pop('embeddings')
-            except Exception:
-                self.logger.info(traceback.format_exc())
+            result = self.es.search(
+                index=self.index_name,
+                query=query,
+                source=True,
+                size=limit,
+            )['hits']['hits']
+            doc.matches = ESConverter.convert_es_results_to_matches(
+                query_doc=doc,
+                es_results=result,
+                get_score_breakdown=get_score_breakdown,
+                metric=self.metric,
+                semantic_scores=self.default_semantic_scores,
+            )
+            doc.tags.pop('embeddings')
         return DocumentArray(list(zip(*es_queries))[0])
 
     @secure_request(on='/update', level=SecurityLevel.USER)
@@ -313,6 +308,16 @@ class NOWElasticIndexer(Executor):
                 f"Deleted {resp['deleted']} documents in Elasticsearch index {self.index_name}"
             )
         return DocumentArray()
+
+    def aggregate_embeddings(self, docs_map: Dict[str, DocumentArray]):
+        """Aggregate embeddings of cc level to c level.
+
+        :param docs_map: a dictionary of `DocumentArray`s, where the key is the embedding space aka encoder name.
+        """
+        for docs in docs_map.values():
+            for doc in docs:
+                for c in doc.chunks:
+                    c.embedding = c.chunks.embeddings.mean(axis=0)
 
     @secure_request(on='/tags', level=SecurityLevel.USER)
     def get_tags_and_values(self, **kwargs):
